splash: Fix more compilation warnings on 64-bit platforms

+++ common/splash_source.c
@@ -37,7 +37,7 @@ static int splash_sf_read_raw(uintptr_t bmp_load_addr, int offset, size_t read_s
 	return spi_flash_read(sf, offset, read_size, (void *)bmp_load_addr);
 }
 #else
-static int splash_sf_read_raw(u32 bmp_load_addr, int offset, size_t read_size)
+static int splash_sf_read_raw(uintptr_t bmp_load_addr, int offset, size_t read_size)
 {
 	debug("%s: sf support not available\n", __func__);
 	return -ENOSYS;
@@ -54,7 +54,7 @@ static int splash_nand_read_raw(uintptr_t bmp_load_addr, int offset, size_t read
 				  (u_char *)bmp_load_addr);
 }
 #else
-static int splash_nand_read_raw(u32 bmp_load_addr, int offset, size_t read_size)
+static int splash_nand_read_raw(uintptr_t bmp_load_addr, int offset, size_t read_size)
 {
 	debug("%s: nand support not available\n", __func__);
 	return -ENOSYS;
@@ -326,7 +326,7 @@ static int splash_load_fit(struct splash_location *location, uintptr_t bmp_load_
 
 	/* Read in entire FIT */
 	fit_header = (const u32 *)(bmp_load_addr + header_size);
-	res = splash_storage_read_raw(location, (u32)fit_header, fit_size);
+	res = splash_storage_read_raw(location, (uintptr_t)fit_header, fit_size);
 	if (res < 0)
 		return res;
